@@ -1,8 +1,13 @@
 module boringssl.googlesource.com/boringssl
 
-go 1.13
+go 1.19
 
 require (
 	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
 	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
+)
+
+require (
+	golang.org/x/sys v0.0.0-20210423082822-04245dca01da // indirect
+	golang.org/x/term v0.0.0-20201126162022-7de9c90e9dd1 // indirect
 )
